Move all texture creation functions into texture_util.
PiperOrigin-RevId: 892838123
